refactor(api): move step logic to separate package
