+++ api/build.go
@@ -1677,6 +1677,116 @@ func planServices(database database.Interface, p *pipeline.Build, b *library.Bui
 	return services, nil
 }
 
+// planSteps is a helper function to plan all steps
+// in the build for execution. This creates the steps
+// for the build in the configured backend.
+func planSteps(database database.Interface, p *pipeline.Build, b *library.Build) ([]*library.Step, error) {
+	// variable to store planned steps
+	steps := []*library.Step{}
+
+	// iterate through all pipeline stages
+	for _, stage := range p.Stages {
+		// iterate through all steps for each pipeline stage
+		for _, step := range stage.Steps {
+			// create the step object
+			s := new(library.Step)
+			s.SetBuildID(b.GetID())
+			s.SetRepoID(b.GetRepoID())
+			s.SetNumber(step.Number)
+			s.SetName(step.Name)
+			s.SetImage(step.Image)
+			s.SetStage(stage.Name)
+			s.SetStatus(constants.StatusPending)
+			s.SetCreated(time.Now().UTC().Unix())
+
+			// send API call to create the step
+			err := database.CreateStep(s)
+			if err != nil {
+				return steps, fmt.Errorf("unable to create step %s: %w", s.GetName(), err)
+			}
+
+			// send API call to capture the created step
+			s, err = database.GetStepForBuild(b, s.GetNumber())
+			if err != nil {
+				return steps, fmt.Errorf("unable to get step %s: %w", s.GetName(), err)
+			}
+
+			// populate environment variables from step library
+			//
+			// https://pkg.go.dev/github.com/go-vela/types/library#step.Environment
+			err = step.MergeEnv(s.Environment())
+			if err != nil {
+				return steps, err
+			}
+
+			// create the log object
+			l := new(library.Log)
+			l.SetStepID(s.GetID())
+			l.SetBuildID(b.GetID())
+			l.SetRepoID(b.GetRepoID())
+			l.SetData([]byte{})
+
+			// send API call to create the step logs
+			err = database.CreateLog(l)
+			if err != nil {
+				return nil, fmt.Errorf("unable to create logs for step %s: %w", s.GetName(), err)
+			}
+
+			steps = append(steps, s)
+		}
+	}
+
+	// iterate through all pipeline steps
+	for _, step := range p.Steps {
+		// create the step object
+		s := new(library.Step)
+		s.SetBuildID(b.GetID())
+		s.SetRepoID(b.GetRepoID())
+		s.SetNumber(step.Number)
+		s.SetName(step.Name)
+		s.SetImage(step.Image)
+		s.SetStatus(constants.StatusPending)
+		s.SetCreated(time.Now().UTC().Unix())
+
+		// send API call to create the step
+		err := database.CreateStep(s)
+		if err != nil {
+			return steps, fmt.Errorf("unable to create step %s: %w", s.GetName(), err)
+		}
+
+		// send API call to capture the created step
+		s, err = database.GetStepForBuild(b, s.GetNumber())
+		if err != nil {
+			return steps, fmt.Errorf("unable to get step %s: %w", s.GetName(), err)
+		}
+
+		// populate environment variables from step library
+		//
+		// https://pkg.go.dev/github.com/go-vela/types/library#step.Environment
+		err = step.MergeEnv(s.Environment())
+		if err != nil {
+			return steps, err
+		}
+
+		// create the log object
+		l := new(library.Log)
+		l.SetStepID(s.GetID())
+		l.SetBuildID(b.GetID())
+		l.SetRepoID(b.GetRepoID())
+		l.SetData([]byte{})
+
+		// send API call to create the step logs
+		err = database.CreateLog(l)
+		if err != nil {
+			return steps, fmt.Errorf("unable to create logs for step %s: %w", s.GetName(), err)
+		}
+
+		steps = append(steps, s)
+	}
+
+	return steps, nil
+}
+
 // cleanBuild is a helper function to kill the build
 // without execution. This will kill all resources,
 // like steps and services, for the build in the
