+++ api/step/update.go
@@ -1,0 +1,163 @@
+// Copyright (c) 2023 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+package step
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/go-vela/server/database"
+	"github.com/go-vela/server/router/middleware/build"
+	"github.com/go-vela/server/router/middleware/org"
+	"github.com/go-vela/server/router/middleware/repo"
+	"github.com/go-vela/server/router/middleware/step"
+	"github.com/go-vela/server/router/middleware/user"
+	"github.com/go-vela/server/util"
+	"github.com/go-vela/types/library"
+	"github.com/sirupsen/logrus"
+)
+
+// swagger:operation PUT /api/v1/repos/{org}/{repo}/builds/{build}/steps/{step} steps UpdateStep
+//
+// Update a step for a build
+//
+// ---
+// produces:
+// - application/json
+// parameters:
+// - in: path
+//   name: org
+//   description: Name of the org
+//   required: true
+//   type: string
+// - in: path
+//   name: repo
+//   description: Name of the repo
+//   required: true
+//   type: string
+// - in: path
+//   name: build
+//   description: Build number
+//   required: true
+//   type: integer
+// - in: path
+//   name: step
+//   description: Step number
+//   required: true
+//   type: integer
+// - in: body
+//   name: body
+//   description: Payload containing the step to update
+//   required: true
+//   schema:
+//     "$ref": "#/definitions/Step"
+// security:
+//   - ApiKeyAuth: []
+// responses:
+//   '200':
+//     description: Successfully updated the step
+//     schema:
+//       "$ref": "#/definitions/Step"
+//   '400':
+//     description: Unable to update the step
+//     schema:
+//       "$ref": "#/definitions/Error"
+//   '500':
+//     description: Unable to update the step
+//     schema:
+//       "$ref": "#/definitions/Error"
+
+// UpdateStep represents the API handler to update
+// a step for a build in the configured backend.
+func UpdateStep(c *gin.Context) {
+	// capture middleware values
+	b := build.Retrieve(c)
+	o := org.Retrieve(c)
+	r := repo.Retrieve(c)
+	s := step.Retrieve(c)
+	u := user.Retrieve(c)
+
+	entry := fmt.Sprintf("%s/%d/%d", r.GetFullName(), b.GetNumber(), s.GetNumber())
+
+	// update engine logger with API metadata
+	//
+	// https://pkg.go.dev/github.com/sirupsen/logrus?tab=doc#Entry.WithFields
+	logrus.WithFields(logrus.Fields{
+		"build": b.GetNumber(),
+		"org":   o,
+		"repo":  r.GetName(),
+		"step":  s.GetNumber(),
+		"user":  u.GetName(),
+	}).Infof("updating step %s", entry)
+
+	// capture body from API request
+	input := new(library.Step)
+
+	err := c.Bind(input)
+	if err != nil {
+		retErr := fmt.Errorf("unable to decode JSON for step %s: %w", entry, err)
+
+		util.HandleError(c, http.StatusBadRequest, retErr)
+
+		return
+	}
+
+	// update step fields if provided
+	if len(input.GetStatus()) > 0 {
+		// update status if set
+		s.SetStatus(input.GetStatus())
+	}
+
+	if len(input.GetError()) > 0 {
+		// update error if set
+		s.SetError(input.GetError())
+	}
+
+	if input.GetExitCode() > 0 {
+		// update exit_code if set
+		s.SetExitCode(input.GetExitCode())
+	}
+
+	if input.GetStarted() > 0 {
+		// update started if set
+		s.SetStarted(input.GetStarted())
+	}
+
+	if input.GetFinished() > 0 {
+		// update finished if set
+		s.SetFinished(input.GetFinished())
+	}
+
+	if len(input.GetHost()) > 0 {
+		// update host if set
+		s.SetHost(input.GetHost())
+	}
+
+	if len(input.GetRuntime()) > 0 {
+		// update runtime if set
+		s.SetRuntime(input.GetRuntime())
+	}
+
+	if len(input.GetDistribution()) > 0 {
+		// update distribution if set
+		s.SetDistribution(input.GetDistribution())
+	}
+
+	// send API call to update the step
+	err = database.FromContext(c).UpdateStep(s)
+	if err != nil {
+		retErr := fmt.Errorf("unable to update step %s: %w", entry, err)
+
+		util.HandleError(c, http.StatusInternalServerError, retErr)
+
+		return
+	}
+
+	// send API call to capture the updated step
+	s, _ = database.FromContext(c).GetStepForBuild(b, s.GetNumber())
+
+	c.JSON(http.StatusOK, s)
+}
